requirements.txt
# Bot Framework
aiogram==3.3.0
python-dotenv==1.0.0

# Database
SQLAlchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1

# Web Server
aiohttp==3.9.1
gunicorn==21.2.0
uvicorn==0.24.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
Pillow==10.1.0
emoji==2.9.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Documentation
Sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# Security
cryptography==41.0.5
python-jose==3.3.0
passlib==1.7.4

# Monitoring
prometheus-client==0.19.0
sentry-sdk==1.35.0

# Caching
aioredis==2.0.1
cachetools==5.3.2

# Rate Limiting
aioredis==2.0.1
limits==3.7.0

# Logging
structlog==23.2.0
python-json-logger==2.0.7

# Type Checking
types-redis==4.6.0.20231106
types-Pillow==10.1.0.20231106
types-python-dateutil==2.8.19.20231106
types-pytz==2023.3.0.20231106 